+++ modules/canvaskit/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - GrContexts will properly target the correct WebGL context
  - CanvasKit built with no_embedded_font will properly link and be able to load fonts from passed-in
    bytes.
+ - Text styled with fontSize or heightMultiplier 0 will be invisible.
 
 ## [0.37.0] - 2022-09-07